Add support for folder dependencies.

+++ src/gunilla/actions/download.py
@@ -1,5 +1,6 @@
-from gunilla.config import instance as config_instance
+from gunilla.config import instance as config_instance, DependencyType
 from gunilla.environment import instance as env_instance
+from gunilla.exceptions import ActionException
 import json
 import os
 import requests
@@ -9,24 +10,32 @@ from zipfile import ZipFile
 
 def run():
     plugin_dependencies = config_instance().dependencies.plugins
-    for slug in plugin_dependencies:
-        dependency = plugin_dependencies[slug]
-        version = dependency.version
-        print("Downloading plugin '{}' at version '{}'".format(slug, version))
-
-        descriptor = json.loads(requests.get('https://api.wordpress.org/plugins/info/1.0/{}.json'.format(slug)).text)
-        if env_instance().debug:
-            print("Downloaded descriptor: %s" % json.dumps(descriptor, indent=2))
-        download_extract(slug, dependency, descriptor, 'plugins')
+    copy_dependencies(plugin_dependencies, 'https://api.wordpress.org/plugins/info/1.0/{}.json', 'plugins')
 
     theme_dependencies = config_instance().dependencies.themes
-    for slug in theme_dependencies:
-        dependency = theme_dependencies[slug]
-        version = dependency.version
-        print("Downloading theme '{}' at version '{}'".format(slug, version))
+    copy_dependencies(theme_dependencies, 'https://api.wordpress.org/themes/info/1.1/?action=theme_information&request[slug]={}&request[fields][versions]=true', 'themes')
 
-        descriptor = json.loads(requests.get('https://api.wordpress.org/themes/info/1.1/?action=theme_information&request[slug]={}&request[fields][versions]=true'.format(slug)).text)
-        download_extract(slug, dependency, descriptor, 'themes')
+
+def copy_dependencies(dependencies, url_template, folder):
+    for slug in dependencies:
+        dependency = dependencies[slug]
+        if dependency.type == DependencyType.DOWNLOAD:
+            download_dependency(dependencies, slug, url_template, folder)
+        elif dependency.type == DependencyType.FOLDER:
+            copy_folder_dependency(dependencies, slug, folder)
+        else:
+            print("Skipping dependency {}".format(slug))
+
+
+def download_dependency(dependencies, slug, url_template, folder):
+    dependency = dependencies[slug]
+    version = dependency.version
+    print("Downloading plugin '{}' at version '{}'".format(slug, version))
+
+    descriptor = json.loads(requests.get(url_template.format(slug)).text)
+    if env_instance().debug:
+        print("Downloaded descriptor: %s" % json.dumps(descriptor, indent=2))
+    download_extract(slug, dependency, descriptor, folder)
 
 
 def download_extract(slug, dependency, descriptor, folder):
@@ -34,6 +43,17 @@ def download_extract(slug, dependency, descriptor, folder):
     if version == 'latest':
         version = descriptor['version']
         print("Latest version of '{}' is '{}'".format(slug, version))
+
+    if version not in descriptor['versions']:
+        print("Version {} does not seem to be available".format(version))
+        print("Available versions are:")
+        sorted_versions = []
+        for key in descriptor['versions']:
+            sorted_versions.append(key)
+        sorted_versions.sort()
+        for version in sorted_versions:
+            print(version)
+        raise ActionException("Provided version is not available")
 
     download_request = requests.get(descriptor['versions'][version])
     shutil.rmtree('{}/{}'.format(folder, slug), True)
@@ -45,3 +65,12 @@ def download_extract(slug, dependency, descriptor, folder):
     zip_file = ZipFile(zip_file_name)
     zip_file.extractall('{}'.format(folder))
     os.remove(zip_file_name)
+
+def copy_folder_dependency(dependencies, slug, folder):
+    dependency = dependencies[slug]
+    src_folder = dependency['path']
+    dest_folder = os.path.join(folder, slug)
+
+    print("Copying plugin '{}' files from folder '{}'".format(slug, src_folder))
+    shutil.rmtree(dest_folder, True)
+    shutil.copytree(src_folder, dest_folder, True, None)
